+++ src/components/AttachmentList.vue
@@ -199,7 +199,7 @@ export default {
 }
 
 .attachment:hover {
-  color: #333;
+  color: #000;
   text-decoration: none;
 }