+++ src/components/Hub/HubCard.vue
@@ -32,7 +32,7 @@
     >
     <v-card-text
       v-if="item.authors"
-      style="font-weight: 700; color: #333; font-size: 12px; margin-top: -40px"
+      style="font-weight: 700; color: #000; font-size: 12px; margin-top: -40px"
       >{{ displayAuthors(item.authors) }}</v-card-text
     >
 
@@ -321,7 +321,7 @@ export default {
   cursor: pointer;
 }
 .category:hover {
-  color: #444;
+  color: #000;
   text-decoration: underline;
 }
 </style>