+++ src/views/Search.vue
@@ -204,6 +204,6 @@ export default {
   color: #222;
 }
 .author:hover {
-  color: #333;
+  color: #000;
 }
 </style>